Banner set to be displayed before user login.
