+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/client/ui/login/LoginUi.ui.xml
@@ -42,6 +42,22 @@
 				
 		</b:Modal>
 		</g:FormPanel>
+        
+        <b:Modal closable="true" fade="true" dataBackdrop="STATIC"
+            dataKeyboard="true" ui:field="accessBannerModal"
+            b:id="accessBannerModal">
+            <b:ModalBody>
+                <b:Panel>
+                    <b:PanelBody>
+                        <b.html:Strong ui:field="accessBannerModalPannelBody"/>
+                    </b:PanelBody>
+                </b:Panel>
+            </b:ModalBody>
+            <b:ModalFooter>
+                <b:Button type="PRIMARY" dataDismiss="MODAL"
+                    ui:field="buttonAccessBannerModalOk" />
+            </b:ModalFooter>
+        </b:Modal>
 
 
 		<kura:AlertDialog ui:field="alertDialog" />
